final version

+++ delete.php
@@ -10,7 +10,7 @@ if(!isset($_SESSION["user"])){
 
 $id = $_GET["id"];
 
-$statement = $pdo->prepare("SELECT * FROM contacts WHERE id = :id");
+$statement = $pdo->prepare("SELECT * FROM contacts WHERE id = :id LIMIT 1");
 $statement->execute([":id" => $id]);
 
 if ($statement->rowCount() == 0) {
@@ -19,8 +19,15 @@ if ($statement->rowCount() == 0) {
     return;
 }
 
-
-$pdo->prepare("DELETE FROM contacts WHERE id = :id")->execute([":id" => $id]);
+$contact = $statement->fetch(PDO::FETCH_ASSOC);
+if($contact["user_id"] != $_SESSION["user"]["id"]){
+    http_response_code(403);
+    echo ("HTTP 403 UNAUTORIZED PERRA");
+    return;
+} else {
+    $pdo->prepare("DELETE FROM contacts WHERE id = :id")->execute([":id" => $id]);
+}
+$_SESSION["flash"] = ["message" => "Contact {$contact['name']} deleted."];
 
 header("Location: index.php");
